fix(types): ship self-contained declarations; split types by concern
