+++ lib/components/FormSelectField.vue
@@ -47,9 +47,9 @@
 
 <script lang="ts" setup>
 import { computed } from "vue";
-import type { Formik } from "@/types";
-import useFormikContext from "@/composables/useFormikContext";
-import { constructLabel } from "@/helpers";
+import type { Formik } from "../types";
+import useFormikContext from "../composables/useFormikContext";
+import { constructLabel } from "../helpers";
 
 const props = defineProps<{
   formik?: Formik;
