+++ lib/composables/useFieldArray.ts
@@ -1,4 +1,4 @@
-import type { Formik } from "@/types";
+import type { Formik } from "../types";
 
 function useFieldArray(formik: Formik) {
   /**
