+++ lib/helpers/index.ts
@@ -1,10 +1,10 @@
-import setNestedValue from "@/helpers/formikObject/setNestedValue";
-import getNestedValue from "@/helpers/formikObject/getNestedValue";
-import clearObject from "@/helpers/formikObject/clearObject";
-import deepEqual from "@/helpers/formikObject/deepEqual";
-import debounce from "@/helpers/formikObject/debounce";
-import { constructLabel } from "@/helpers/utils";
-import { applyState, clearState, assignClonedState } from "@/helpers/formikObject/stateManagement";
+import setNestedValue from "./formikObject/setNestedValue";
+import getNestedValue from "./formikObject/getNestedValue";
+import clearObject from "./formikObject/clearObject";
+import deepEqual from "./formikObject/deepEqual";
+import debounce from "./formikObject/debounce";
+import { constructLabel } from "./utils";
+import { applyState, clearState, assignClonedState } from "./formikObject/stateManagement";
 
 export {
   setNestedValue,
