+++ lib/index.ts
@@ -1,10 +1,10 @@
-import useFormik from "@/composables/useFormik";
-import useFieldArray from "@/composables/useFieldArray";
-import FormTextArea from "@/components/FormTextArea.vue";
-import FormInput from "@/components/FormInput.vue";
-import FormSelectField from "@/components/FormSelectField.vue";
-import FormContentEditable from "@/components/FormContentEditable.vue";
-import FormikForm from "@/components/FormikForm.vue";
+import useFormik from "./composables/useFormik";
+import useFieldArray from "./composables/useFieldArray";
+import FormTextArea from "./components/FormTextArea.vue";
+import FormInput from "./components/FormInput.vue";
+import FormSelectField from "./components/FormSelectField.vue";
+import FormContentEditable from "./components/FormContentEditable.vue";
+import FormikForm from "./components/FormikForm.vue";
 
 export {
   useFormik,
