+++ lib/types/common.ts
@@ -1,0 +1,4 @@
+/* eslint-disable-next-line @typescript-eslint/no-explicit-any */
+export type AllowedAny = any;
+
+export type MaybePromise<T> = T | Promise<T>;
