+++ package.json
@@ -68,38 +68,38 @@
     "node": ">=16.0.0"
   },
   "devDependencies": {
-    "@tsconfig/node22": "^22.0.0",
-    "@types/jsdom": "^27.0.0",
+    "@tsconfig/node22": "^22.0.5",
+    "@types/jsdom": "^28.0.3",
     "@types/lodash.clonedeep": "^4.5.9",
-    "@types/node": "^24.9.0",
+    "@types/node": "^26.0.0",
     "@vitejs/plugin-vue": "^6.0.1",
     "@vitejs/plugin-vue-jsx": "^5.1.1",
-    "@vitest/coverage-istanbul": "^4.0.18",
-    "@vitest/eslint-plugin": "1.4.0",
-    "@vitest/ui": "^4.0.18",
+    "@vitest/coverage-istanbul": "^4.1.9",
+    "@vitest/eslint-plugin": "^1.6.20",
+    "@vitest/ui": "^4.1.9",
     "@vue/eslint-config-prettier": "^10.2.0",
-    "@vue/eslint-config-typescript": "^14.6.0",
-    "@vue/test-utils": "^2.4.6",
-    "@vue/tsconfig": "^0.8.0",
-    "eslint": "^9.39.0",
-    "eslint-plugin-vue": "^10.5.1",
+    "@vue/eslint-config-typescript": "^14.9.0",
+    "@vue/test-utils": "^2.4.11",
+    "@vue/tsconfig": "^0.9.0",
+    "eslint": "^10.5.0",
+    "eslint-plugin-vue": "^10.9.2",
     "husky": "^9.1.7",
-    "joi": "^18.0.1",
-    "jsdom": "^27.1.0",
-    "npm-run-all2": "^8.0.4",
-    "prettier": "^3.6.2",
-    "sass-embedded": "^1.93.3",
+    "joi": "^18.2.3",
+    "jsdom": "^29.1.1",
+    "lodash.clonedeep": "^4.5.0",
+    "npm-run-all2": "^9.0.2",
+    "prettier": "^3.8.4",
+    "sass-embedded": "^1.100.0",
     "superstruct": "^2.0.2",
-    "typescript": "^5.9.3",
-    "vite": "^7.1.0",
-    "vite-plugin-dts": "^4.5.0",
-    "vite-plugin-vue-devtools": "^8.0.3",
-    "vitest": "^4.0.18",
-    "vue-tsc": "^3.1.2",
+    "typescript": "^6.0.3",
+    "vite": "^8.1.0",
+    "vite-plugin-dts": "^5.0.3",
+    "vite-plugin-vue-devtools": "^8.1.3",
+    "vitest": "^4.1.9",
+    "vue": "^3.5.38",
+    "vue-tsc": "^3.3.5",
     "yup": "^1.7.1",
-    "zod": "^4.1.0",
-    "vue": "^3.5.13",
-    "lodash.clonedeep": "^4.5.0"
+    "zod": "^4.4.3"
   },
   "license": "LGPL-3.0-or-later",
   "pre-commit": {
