+++ vitest.config.ts
@@ -2,8 +2,12 @@ import { fileURLToPath } from 'node:url'
 import { mergeConfig, defineConfig } from 'vitest/config'
 import viteConfig from './vite.config'
 
+// vite.config exports a config function (to gate dev-only plugins by command);
+// resolve it to a plain config object before merging.
+const resolvedViteConfig = viteConfig({ command: 'serve', mode: 'test' })
+
 export default mergeConfig(
-  viteConfig,
+  resolvedViteConfig,
   defineConfig({
     test: {
       environment: 'jsdom',
